Created placeBet function, started on validators

+++ database/migrations/2022_08_22_161644_create_bets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('bets', function (Blueprint $table) {
             $table->id();
-            $table->integer('stake_amount');
+            $table->decimal('stake_amount',5);
             $table->timestamps();
         });
     }
